requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
beautifulsoup4==4.12.3
blinker==1.9.0
catboost==1.2.7
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cssutils==2.11.1
currency-symbols==2.0.3
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
Flask==3.0.3
fonttools==4.56.0
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lightgbm==4.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
narwhals==1.29.0
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.5
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.11.1
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.33.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.1
threadpoolctl==3.5.0
tornado==6.4.2
traitlets==5.14.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wordcloud==1.9.4
xgboost==2.1.4
xlrd==2.0.1
xls2xlsx==0.2.0
